+++ src/datatable/HISTORY.md
@@ -3,6 +3,10 @@ DataTable Change History
 
 @VERSION@
 ------
+* Add highlight module [Pull Request #1196]
+
+* Document updates and variable changes to improve understanding of code
+  [Pull Request #946] [Satyam]
 
 * Add Show All to language packs. [Pull Request #1173] [Issue #1167]